kie-issues#91: refactoring how/where extensions are dealt with
