+++ packages/serverless-logic-web-tools/src/editor/EditorToolbar.tsx
@@ -60,7 +60,7 @@ import * as React from "react";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useHistory } from "react-router";
 import { Alerts, AlertsController, useAlert } from "../alerts/Alerts";
-import { isDashbuilder, isServerlessWorkflow } from "../extension";
+import { isDashbuilder, isServerlessWorkflow } from "@kie-tools-core/workspaces-git-fs/dist/constants/ExtensionHelper";
 import { useAppI18n } from "../i18n";
 import {
   useNavigationBlockersBypass,
