+++ packages/serverless-logic-web-tools/src/editor/NewFileDropdownMenu.tsx
@@ -34,7 +34,8 @@ import { Alert, AlertActionCloseButton } from "@patternfly/react-core/dist/js/co
 import { basename } from "path";
 import { ImportFromUrlForm } from "../workspace/components/ImportFromUrlForm";
 import { useRoutes } from "../navigation/Hooks";
-import { FileTypes, isEditable, SupportedFileExtensions } from "../extension";
+import { isEditable, SupportedFileExtensions } from "../extension";
+import { FileTypes } from "@kie-tools-core/workspaces-git-fs/dist/constants/ExtensionHelper";
 import { decoder } from "@kie-tools-core/workspaces-git-fs/dist/encoderdecoder/EncoderDecoder";
 import { extractExtension } from "@kie-tools-core/workspaces-git-fs/dist/relativePath/WorkspaceFileRelativePathParser";
 import { UrlType } from "../workspace/hooks/ImportableUrlHooks";
