+++ packages/serverless-logic-web-tools/src/editor/api/SwfServiceCatalogStore.ts
@@ -23,7 +23,7 @@ import { parseApiContent } from "@kie-tools/serverless-workflow-service-catalog/
 import { ArtifactType, SearchedArtifact } from "@rhoas/registry-instance-sdk";
 import { OpenAPIV3 } from "openapi-types";
 import * as yaml from "yaml";
-import { isSpec } from "../../extension";
+import { isSpec } from "@kie-tools-core/workspaces-git-fs/dist/constants/ExtensionHelper";
 import {
   isServiceAccountConfigValid,
   ServiceAccountSettingsConfig,
