+++ packages/serverless-logic-web-tools/src/home/SampleCard.tsx
@@ -9,7 +9,7 @@ import { Text } from "@patternfly/react-core/dist/js/components/Text";
 import { Label, LabelProps } from "@patternfly/react-core/dist/js/components/Label";
 import { FolderIcon, FileIcon, MonitoringIcon } from "@patternfly/react-icons/dist/js/icons";
 import { labelColors } from "../workspace/components/FileLabel";
-import { FileTypes } from "../extension";
+import { FileTypes } from "@kie-tools-core/workspaces-git-fs/dist/constants/ExtensionHelper";
 
 export enum SampleType {
   SW_YML = "sw.yml",
