+++ packages/serverless-logic-web-tools/src/virtualServiceRegistry/hooks/useVirtualServiceRegistryDependencies.tsx
@@ -18,7 +18,7 @@ import * as React from "react";
 import { useWorkspaces, WorkspaceFile } from "@kie-tools-core/workspaces-git-fs/dist/context/WorkspacesContext";
 import { ActiveWorkspace } from "@kie-tools-core/workspaces-git-fs/dist/model/ActiveWorkspace";
 import { useState, useEffect, useMemo } from "react";
-import { isServerlessWorkflow } from "../../extension";
+import { isServerlessWorkflow } from "@kie-tools-core/workspaces-git-fs/dist/constants/ExtensionHelper";
 import { getVirtualServiceRegistryDependencies } from "../models/VirtualServiceRegistryFunction";
 import { useVirtualServiceRegistry } from "../VirtualServiceRegistryContext";
